% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedDotPlot.R
\name{AnnotatedDotPlot}
\alias{AnnotatedDotPlot}
\title{AnnotatedDotPlot for Seurat or post-integration (pi) object}
\usage{
AnnotatedDotPlot(
  object,
  features,
  cols = c("lightgrey", "blue"),
  split.by = NULL,
  group.by = NULL,
  clust.row = TRUE,
  clust.column = FALSE,
  column.annotation = NULL,
  column.annotation.cols = NULL,
  column.fontsize = 12,
  row.fontsize = 12,
  legend.label.fontsize = 13,
  legend.title.fontsize = 15,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{Genes/features to plot}

\item{cols}{Colors to plot. Same as the "cols" argument of \code{\link[Seurat]{DotPlot}} (default: c("lightgrey", "blue"))}

\item{split.by}{Factor to split the groups by. Same as the "split.by" argument of \code{\link[Seurat]{DotPlot}}}

\item{group.by}{Factor to group the cells by. Same as the "group.by" argument of \code{\link[Seurat]{DotPlot}}}

\item{clust.row}{Whether to cluster the rows/identities (default: TRUE)}

\item{clust.column}{Whether to cluster the columns/features (default: FALSE)}

\item{column.annotation}{A data.frame contains features and their annotation groups}

\item{column.annotation.cols}{Colors for column annotation}

\item{column.fontsize}{Size of column text (default: 12)}

\item{row.fontsize}{Size of row text (default: 12)}

\item{legend.label.fontsize}{Size of the legend labels (default: 13)}

\item{legend.title.fontsize}{Size of the legend title (default: 15)}

\item{...}{Extra parameters passed to DotPlot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
The AnnotatedDotPlot function enhances the original \code{\link[Seurat]{DotPlot}} function from Seurat by:
(1) adding feature annotations to the columns); (2) allowing clustering and visualzation of dendrograms of
identities on the rows. Use its wrapper function \code{\link[Ragas]{RunAnnotatedDotPlot}} for plotting.
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
\keyword{internal}
