% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{CheckPiData}
\alias{CheckPiData}
\alias{CheckPiData.default}
\title{Check Validity for PiData}
\usage{
CheckPiData(x, ...)

\method{CheckPiData}{default}(x, seurat.obj, verbose = FALSE)
}
\arguments{
\item{x}{A PiData object}

\item{...}{Additional arguments}

\item{seurat.obj}{A Seurat object}

\item{verbose}{Whether to print messages (default: FALSE)}
}
\description{
Check the validity of PiData/PiDataList/Pi object
}
\seealso{
Usage of \code{CheckPiData} for different classes: \code{\link[Ragas]{PiExpFreqData}}, \code{\link[Ragas]{PiMarkerData}},
\code{\link[Ragas]{PiDSData}}, \code{\link[Ragas]{PiCellPropData}}, \code{\link[Ragas]{PiParentMetaData}},
\code{\link[Ragas]{PiDataList}}, \code{\link[Ragas]{Pi}}.
}
