% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConfigureReprojection}
\alias{ConfigureReprojection}
\title{Create UMAP configuration for re-projection}
\usage{
ConfigureReprojection(
  type = c("main", "sc"),
  sc.name = NULL,
  umap.name = "umap",
  umap.input.reduction.name = "harmony",
  umap.dims = 1:20,
  umap.n.neighbors = 30,
  umap.metric = "cosine",
  append.to = NULL
)
}
\arguments{
\item{type}{Type of UMAP configure, either "main" for main cluster or "sc" for subcluster}

\item{sc.name}{Name of the child object for subcluster analysis, which should be consistent with the \code{child.object.list}
argument of the \code{\link[Ragas]{CreatePostIntegrationObject}} function}

\item{umap.name}{Name of the UMAP object (default: umap)}

\item{umap.input.reduction.name}{Name of the reduction as input to re-run UMAP (default: harmony)}

\item{umap.dims}{Input dimensions to run UMAP with. Same as the \code{dims} argument of the \code{\link[Seurat]{RunUMAP}} function (default: 1:20).}

\item{umap.n.neighbors}{Number of neighbors for the KNN graph. Same as the \code{n.neighbors} argument of the \code{\link[Seurat]{RunUMAP}} function (default: 30)}

\item{umap.metric}{Distance metric. Same as the \code{metric} argument of the \code{\link[Seurat]{RunUMAP}} function (default: "cosine")}

\item{append.to}{Existing subcluster configuration to append to}
}
\value{
A list containing configurations for UAMP re-projection
}
\description{
Allow users to provide custom parameters to retrieve KNN data or re-run UMAP for re-projection
}
\examples{
\dontrun{
my.config <- ConfigureReprojection(type = "sc", sc.name = "Bcell")
my.config <- ConfigureReprojection(type = "sc", sc.name = "Tcell", append.to = my.config)
}

}
