% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.bcell.pi}
\alias{csle.bcell.pi}
\title{An example Pi object built from csle.bcell.small}
\format{
A \code{\link[Ragas]{Pi}} object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.bcell.pi
}
\description{
This data contained down-sampled B cells from Nehar-Belaid et al., 2020.
Run the example code to re-produce the Pi object from Seurat object \code{\link[Ragas]{csle.bcell.small}}:
}
\examples{
\dontrun{
csle.bcell.pi <- csle.bcell.small \%>\%
CreatePostIntegrationObject(parent.object = csle.pbmc.small) \%>\%
CalculateExpFreqs(ident = "cluster.annotation") \%>\%
RunFindAllMarkers(ident = "cluster.annotation") \%>\%
RunPseudobulkAnalysis(ident.var = "cluster.annotation",
                      group.var = "Groups",
                      sample.var = "Names",
                      group.1 = "cSLE",
                      group.2 = "cHD") \%>\%
RunProportionPlot(group.by = "Groups",
                  ident = "cluster.annotation",
                  method = "unpooled",
                  unpool.by = "Names",
                  use.parent.as.ref = TRUE
                  )
}
}
\keyword{datasets}
