% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.cd4.mem.pi}
\alias{csle.cd4.mem.pi}
\title{An example Pi object built from csle.cd4.mem.small}
\format{
A \code{\link[Ragas]{Pi}} object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.cd4.mem.pi
}
\description{
This data contained down-sampled CD4+ memory T cells from Nehar-Belaid et al., 2020.
Run the example code to re-produce the Pi object from Seurat object \code{\link[Ragas]{csle.cd4.mem.small}}:
}
\examples{
\dontrun{
csle.cd4.mem.pi <- csle.cd4.mem.small \%>\%
CreatePostIntegrationObject(parent.object = csle.tcell.small) \%>\% ## use total T cell as reference for later proportion analysis
CalculateExpFreqs(ident = "cluster.annotation") \%>\%
RunFindAllMarkers(ident = "cluster.annotation") \%>\%
RunPseudobulkAnalysis(ident.var = "cluster.annotation",
                      group.var = "Groups",
                      sample.var = "Names",
                      group.1 = "cSLE",
                      group.2 = "cHD") \%>\%
RunProportionPlot(group.by = "Groups",
                  method = "unpooled",
                  unpool.by = "Names",
                  use.parent.as.ref = TRUE
                  )

}
}
\keyword{datasets}
