% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.cd4.mem.small}
\alias{csle.cd4.mem.small}
\title{A small example of CD4+ memory T cell data from child SLE}
\format{
A Seurat object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.cd4.mem.small
}
\description{
This object contains down-sampled cells from Nehar-Belaid et al., 2020
}
\keyword{datasets}
