% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.pbmc.pi}
\alias{csle.pbmc.pi}
\title{An example Pi object built from csle.pbmc.small}
\format{
A \code{\link[Ragas]{Pi}} object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.pbmc.pi
}
\description{
This data contained 3000 down-sampled PBMC cells from Nehar-Belaid et al., 2020.
Run the example code to re-produce the Pi object from Seurat object \code{\link[Ragas]{csle.pbmc.small}}:
}
\examples{
\dontrun{
## A list of subclusters to combine
subclusters <- list("B cell" = "csle.bcell.small", "T cell" = "csle.tcell.pi")

## UMAP configuration
sc.config <- ConfigureReprojection(type = "sc", "B cell")
sc.config <- ConfigureReprojection(type = "sc", "T cell", umap.name = "rp", append.to = sc.config)

## select columns for cell identities
sc.colnames <- c("B cell" = "cluster.annotation", "T cell" = "subcluster_idents")

## integrate subclusters
library(dplyr)
csle.pbmc.pi <- csle.pbmc.small \%>\%
CreatePostIntegrationObject(child.object.list = subclusters,
                            rp.subcluster.umap.config = sc.config,
                            rp.main.cluster.anno = "cluster.annotation",
                            rp.subcluster.colname = sc.colnames) \%>\%
CalculateExpFreqs(ident = "subcluster_idents") \%>\%
RunFindAllMarkers(ident = "subcluster_idents") \%>\%
RunPseudobulkAnalysis(ident.var = "subcluster_idents",
                      group.var = "Groups",
                      sample.var = "Names",
                      group.1 = "cSLE",
                      group.2 = "cHD") \%>\%
RunProportionPlot(group.by = "Groups",
                  ident = "subcluster_idents",
                  method = "unpooled",
                  unpool.by = "Names"
                  )

}
}
\keyword{datasets}
