% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.pbmc.small}
\alias{csle.pbmc.small}
\title{A small example of PBMC data from child SLE}
\format{
A Seurat object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.pbmc.small
}
\description{
This object contains 3000 down-sampled PBMC cells from Nehar-Belaid et al., 2020
}
\keyword{datasets}
