% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.tcell.pi}
\alias{csle.tcell.pi}
\title{An example Pi object built from csle.tcell.small}
\format{
A \code{\link[Ragas]{Pi}} object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.tcell.pi
}
\description{
This data contained down-sampled T cells from Nehar-Belaid et al., 2020.
Run the example code to re-produce the Pi object from Seurat object \code{\link[Ragas]{csle.tcell.small}}:
}
\examples{
\dontrun{
## A list of T cell subclusters to combine
subclusters <- list("CD4mem" = "csle.cd4.mem.small", "Treg" = "csle.treg.small")

csle.tcell.pi <- csle.tcell.small \%>\%
CreatePostIntegrationObject(child.object.list = subclusters,
                            rp.main.cluster.anno = "cluster.annotation",
                            rp.subcluster.colname = "cluster.annotation",
                            parent.object = csle.pbmc.small) \%>\%
CalculateExpFreqs(ident = "subcluster_idents") \%>\%
RunFindAllMarkers(ident = "subcluster_idents") \%>\%
RunPseudobulkAnalysis(ident.var = "subcluster_idents",
                      group.var = "Groups",
                      sample.var = "Names",
                      group.1 = "cSLE",
                      group.2 = "cHD") \%>\%
RunProportionPlot(ident = "subcluster_idents",
                  group.by = "Groups",
                  method = "unpooled",
                  unpool.by = "Names",
                  use.parent.as.ref = TRUE
                  )

}
}
\keyword{datasets}
