% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csle.treg.small}
\alias{csle.treg.small}
\title{A small example of Treg data from child SLE}
\format{
A Seurat object
}
\source{
\url{https://www.nature.com/articles/s41590-020-0743-0}
}
\usage{
csle.treg.small
}
\description{
This object contains down-sampled cells from Nehar-Belaid et al., 2020
}
\keyword{datasets}
