% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSDotPlot.R
\name{DSDotPlot}
\alias{DSDotPlot}
\title{DotPlot for Differential State Analysis}
\usage{
DSDotPlot(
  object,
  res,
  exp.freq,
  cpm.dat,
  features = NULL,
  p.threshold = 0.05,
  FC.threshold = 1.5,
  n.deg = Inf,
  p.adj.filter = TRUE,
  pct.threshold = 0.1,
  direction = "both",
  exp.filter = TRUE,
  exp.filter.thr = 0,
  group.col.name = NULL,
  sample.col.name = NULL,
  case.group = NULL,
  ref.group = NULL,
  bin.max = 0.5,
  bin.len = 0.1,
  lfc.min = -2,
  lfc.max = 2,
  clust.row = TRUE,
  clust.column = FALSE,
  heatmap.cols = NULL,
  highlight.deg = TRUE,
  deg.col = NULL,
  column.fontsize = 12,
  row.fontsize = 12,
  column.names.rotation = 90,
  axis.text.hjust = 0.5,
  axis.text.vjust = 0.5,
  legend.label.fontsize = 13,
  legend.title.fontsize = 15
)
}
\arguments{
\item{object}{A Seurat object}

\item{res}{A list containing data.frame  with differential testing results. See \code{\link[muscat]{pbDS}}}

\item{exp.freq}{A list containing per gene expression frequencies. See \code{\link[Ragas]{CalculateExpFreqs}}}

\item{cpm.dat}{A data frame containing CPM}

\item{p.threshold}{P-Value cutoff}

\item{FC.threshold}{Fold-change cutoff}

\item{n.deg}{Number of genes to plot}

\item{p.adj.filter}{Whether to apply multiple testing correction}

\item{pct.threshold}{Cutoff for per gene expression frequencies}

\item{direction}{Plot for either up or down-regulated genes, or both}

\item{exp.filter}{Whether to perform additional expression filter based on median expression difference between groups (default: TRUE)}

\item{exp.filter.thr}{Threshold for difference between median expression from two groups being compared (default: 0)}

\item{group.col.name, sample.col.name}{Column names from the Seurat metadata used as group or sample variable by \code{\link[Ragas]{RunPseudobulkAnalysis}}}

\item{case.group, ref.group}{Names for the case and control/reference group, corresponding to group.1 and group.2 from \code{\link[Ragas]{RunPseudobulkAnalysis}}, respectively}

\item{bin.max}{Maximum exp proportion for binning pct expression data (default: 0.5)}

\item{bin.len}{Step size for binning (default: 0.1)}

\item{lfc.min, lfc.max}{Minimum/maximum logFC to plot}

\item{clust.row, clust.column}{Whether to cluster row and columns}

\item{heatmap.cols}{Three colors for heatmap (low, medium, high)}

\item{highlight.deg}{Whether to highlight significant genes}

\item{deg.col}{Color of the highlight}

\item{column.fontsize, row.fontsize}{Column and row font size}

\item{column.names.rotation}{Rotation for column names}

\item{axis.text.hjust, axis.text.vjust}{Horizontal/vertical justification (in [0, 1]) for axis text (default: 0.5)}

\item{legend.label.fontsize, legend.title.fontsize}{Font sizes for legend label and title}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
An internal function that visualizes differential state analysis results from \code{\link[Ragas]{RunPseudobulkAnalysis}}, which also
incorporates per gene expression frequencies calculated from \code{\link[Ragas]{CalculateExpFreqs}} as dot size.
Use its wrapper function \code{\link[Ragas]{RunDSDotPlot}} for plotting.
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
\keyword{internal}
