% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpressionPlot.R
\name{ExpressionPlot}
\alias{ExpressionPlot}
\title{Expression Plot}
\usage{
ExpressionPlot(
  object,
  feature = NULL,
  assay = "RNA",
  group.by = NULL,
  split.by = NULL,
  ident = "seurat_clusters",
  res = NULL,
  plot.ds.p.value = FALSE,
  plot.ds.adjp.value = FALSE,
  colors = NULL,
  random.col.seed = 1,
  point.size = 1,
  axis.text.size = 12,
  axis.text.angle = 90,
  axis.text.hjust = 0,
  axis.text.vjust = 0,
  axis.title.size = 15,
  legend.text.size = 12,
  legend.title.size = 15,
  sig.size = 8,
  sig.bracket.color = "gray50",
  sig.color = "gray50",
  sig.bracket.width = 0.5
)
}
\arguments{
\item{object}{A Seurat object}

\item{feature}{Gene name}

\item{assay}{Name of the assay to use (default: "RNA")}

\item{group.by}{Name of metadata column to group cells by}

\item{split.by}{Name of the metadata column to split expression by (optional)}

\item{ident}{Cell identity variable (default: seurat_clusters)}

\item{res}{A list containing data.frame  with differential testing results. See \code{\link[muscat]{pbDS}}}

\item{plot.ds.p.value}{TRUE/FALSE if p-value from differential testing should be plotted (default: FALSE)}

\item{plot.ds.adjp.value}{TRUE/FALSE if adjusted p-value from differential testing should be plotted (default: FALSE)}

\item{colors}{Colors for plotting. Number of colors must be the equal to number of idents or the number of groups if split.by is used (optional)}

\item{random.col.seed}{Set seed to control colors}

\item{point.size}{Point size for box plot}

\item{axis.text.size, legend.text.size}{Text size for axis or legend (default: 12)}

\item{axis.text.angle}{Rotation of axis text}

\item{axis.text.hjust}{Horizontal justification (in [0, 1]) for axis text}

\item{axis.text.vjust}{Vertical justification (in [0, 1]) for axis text}

\item{axis.title.size, legend.title.size}{Title size for axis or legend (default: 15)}

\item{sig.size}{Asterisk size for significant tests}

\item{sig.bracket.color}{Bracket color for significant tests}

\item{sig.color}{Asterisk color for significant tests}

\item{sig.bracket.width}{Bracket width for significant tests}
}
\description{
An internal function to plot expression of a gene groups by identity class or additional meta data. Use its wrapper function \code{\link[Ragas]{RunExpressionPlot}} for plotting.
}
\keyword{internal}
