% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixPlot.R
\name{MatrixPlot}
\alias{MatrixPlot}
\title{Matrix Plot}
\usage{
MatrixPlot(
  object,
  markers.list,
  n.genes = 3,
  type = "all",
  up.genes = TRUE,
  down.genes = FALSE,
  heatmap.cols = NULL,
  min.exp = -2,
  max.exp = 2,
  column.fontsize = 12,
  row.fontsize = 12,
  column.anno.cols = NULL,
  column.anno.name.fontsize = 15,
  column.anno.name.rot = 45,
  column.anno.name.just = "center",
  legend.label.fontsize = 13,
  legend.title.fontsize = 15,
  heatmap.width = 15,
  heatmap.height = 10
)
}
\arguments{
\item{object}{A Seurat object}

\item{markers.list}{A data.frame that contains marker data from \code{\link[Seurat]{FindAllMarkers}}}

\item{n.genes}{Number of genes to plot for each identity group}

\item{type}{The type of markers, either "all" or "conserved", which correspond results from  \code{\link[Seurat]{FindAllMarkers}}
and \code{\link[Seurat]{FindConservedMarkers}}, respectively. In the current version, only \code{FindAllMarkers} is supported (default: "all")}

\item{up.genes}{Whether to plot the up-regulated genes (default: TRUE)}

\item{down.genes}{Whether to plot the down-regulated genes (default: FALSE)}

\item{heatmap.cols}{Colors to plot the heatmap (optional). A character vector with colors for low, medium and high expression.}

\item{min.exp}{The minimum expression value to plot (default: -2)}

\item{max.exp}{The maximum expression value to plot (default: 2)}

\item{column.fontsize}{Font size for column texts (default: 12)}

\item{row.fontsize}{Font size for row texts (default: 12)}

\item{column.anno.cols}{Named character vector with colors for column annotations (default: NULL). Names should match Idents of object.}

\item{column.anno.name.fontsize}{Font size for column annotations, such as clusters (default: 15)}

\item{column.anno.name.rot}{Rotation of column annotations (default: 45)}

\item{column.anno.name.just}{Alignment of column annotations (default: "center")}

\item{legend.label.fontsize}{Font size for legend label (default: 13)}

\item{legend.title.fontsize}{Font size for legend title (15)}

\item{heatmap.width}{Width of the heatmap (default: 15)}

\item{heatmap.height}{Height of the heatmap (default: 10)}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
\description{
An internal function to make matrix plot based on Seurat object.
}
