% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{Pi}
\alias{Pi}
\alias{CheckPiData.Pi}
\alias{print.Pi}
\alias{[[.Pi}
\alias{[[<-.Pi}
\alias{AddPiData.Pi}
\title{The Post Integration (Pi) Class.}
\usage{
Pi(
  object,
  exp.freq = NULL,
  markers = NULL,
  ds = NULL,
  cell.prop = NULL,
  parent.meta.data = NULL
)

\method{CheckPiData}{Pi}(x, verbose = FALSE, ...)

\method{print}{Pi}(x, ...)

\method{[[}{Pi}(x, i)

\method{[[}{Pi}(x, i) <- value

\method{AddPiData}{Pi}(x, value)
}
\arguments{
\item{object}{A Seurat object}

\item{exp.freq}{A \code{\link[Ragas]{PiDataList}} containing calculated per gene expression frequency}

\item{markers}{A \code{\link[Ragas]{PiDataList}} containing expression markers}

\item{ds}{A \code{\link[Ragas]{PiDataList}} containing differential state analysis results by pseudobulk method}

\item{cell.prop}{A \code{\link[Ragas]{PiDataList}} containing differential analysis results on cell proportions}

\item{parent.meta.data}{A \code{\link[Ragas]{PiDataList}} containing Parent metadata}

\item{x}{A Pi object}

\item{verbose}{Whether to print messages (default: FALSE)}

\item{...}{Additional arguments}

\item{i}{index for list element to extract or replace}

\item{value}{An object of class \code{PiData}, \code{PiDataList}, or \code{Seurat} to replace its current value}

\item{seurat.obj}{A Seurat object}
}
\description{
An object of class Pi is a list of six fields that covers a variety of analyses for single cell RNA-seq.
These six fields are: (1) "seurat.obj", a \code{\link[Seurat]{Seurat}} object that contains processed data (i.e., data that are filtered, normalized,
batch-integrated, clustered and dimension-reduced); (2) "exp.freq", a list that contains per-gene expression frequency, which is the output of
\code{\link[Ragas]{CalculateExpFreqs}}; (3) "markers", a data.frame that contains results from differential gene expression analysis, such as markers
between different clusters, by running \code{\link[Ragas]{RunFindAllMarkers}}; (4) "ds", differential state analysis results for multi-sample, multi-group study
design by running \code{\link[Ragas]{RunPseudobulkAnalysis}}; (5) "cell.prop", differential cell proportion analysis results from \code{\link[Ragas]{RunProportionPlot}};
(6) "Parent.meta.data", a data.frame containing metadata of from the parent Seurat object.
}
\examples{
## Create a minimum Pi object from a Seurat object
Pi(object = csle.pbmc.small)

\dontrun{
## Check the validity of Pi object
CheckPiData(csle.pbmc.pi, verbose = TRUE)}

## print the Pi object
csle.pbmc.pi

## Get a sub-object
csle.pbmc.pi[["exp.freq"]]
csle.pbmc.pi[["ds"]]

}
\seealso{
\code{\link[Ragas]{CreatePostIntegrationObject}} for a formal way to initialize a Pi object.

Other related classes: \code{\link[Ragas]{PiData}}, \code{\link[Ragas]{PiExpFreqData}}, \code{\link[Ragas]{PiMarkerData}},
\code{\link[Ragas]{PiDSData}}, \code{\link[Ragas]{PiCellPropData}}, \code{\link[Ragas]{PiParentMetaData}}, and
\code{\link[Ragas]{PiDataList}}.
}
