% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{PiCellPropData}
\alias{PiCellPropData}
\alias{CheckPiData.PiCellPropData}
\alias{print.PiCellPropData}
\alias{[[.PiCellPropData}
\title{PiCellPropData Class}
\usage{
PiCellPropData(data, ident, method, group, by, parent = NULL)

\method{CheckPiData}{PiCellPropData}(x, seurat.obj)

\method{print}{PiCellPropData}(x, ...)

\method{[[}{PiCellPropData}(x, i)
}
\arguments{
\item{data}{A data.frame containing test results from \code{\link[Ragas]{RunProportionPlot}}}

\item{ident}{Name for the identity data}

\item{method}{A character string describing the analysis method}

\item{group}{Name of the metadata column to group cells by}

\item{parent}{Name of the parent object}

\item{x}{A PiCellPropData object}

\item{seurat.obj}{A Seurat object}

\item{...}{Additional arguments}

\item{i}{Index for the list element to be extracted}
}
\description{
The PiCellPropData inherts \code{\link[Ragas]{PiData}} class, which stores differential cell proportion analysis results from \code{\link[Ragas]{RunProportionPlot}}.
}
\examples{
\dontrun{
## Run cell proportion analysis on a Seurat object and construct a PiCellPropData object
## In this case, use.parent.as.ref = TRUE is not supported.
## Run ?RunProportionPlot to see how to use parent metadata for proportion analysis
res <- RunProportionPlot(csle.bcell.small,
                         ident = "seurat_clusters",
                         group.by = "Groups",
                         method = "unpooled",
                         unpool.by = "Names")
cp <- PiCellPropData(data = res,
                     ident = "seurat_clusters",
                     method = "unpooled",
                     group = "Groups",
                     parent = NULL)
## Print object
cp

## Check object validity
CheckPiData(cp, seurat.obj = csle.bcell.small)

## Add PiCellPropData object to an existing Pi object
my.pi <- CreatePostIntegrationObject(csle.bcell.small) ## a minimum Pi object
my.pi <- AddPiData(my.pi, cp)
}

}
