% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{PiData}
\alias{PiData}
\alias{print.PiData}
\alias{[[.PiData}
\alias{AddPiData.PiData}
\title{PiData Class}
\usage{
PiData(
  data,
  type = c("un-specified", "exp.freq", "markers", "ds", "cell.prop", "parent.meta.data"),
  ident,
  method = NULL,
  param = NULL
)

\method{print}{PiData}(x, ...)

\method{[[}{PiData}(x, i)

\method{AddPiData}{PiData}(x, value)
}
\arguments{
\item{data}{Analysis data. Depending on the type of the analysis, \code{data} can be a list or data.frame}

\item{type}{A character string indicating the type of the analysis ("un-specified", "exp.freq", "markers", "ds", "cell.prop", or "parent.meta.data")}

\item{ident}{Name for the identity data}

\item{method}{A character string describing the analysis method}

\item{param}{Additional descriptions of the analysis parameters}

\item{x}{A PiData object}

\item{...}{Additional arguments}

\item{i}{Index for the list element to be extracted}

\item{value}{A \code{PiData} object}
}
\description{
PiData class is the most fundamental class for Ragas analysis, which is the parent class for \code{\link[Ragas]{PiExpFreqData}}, \code{\link[Ragas]{PiMarkerData}},
\code{\link[Ragas]{PiDSData}}, \code{\link[Ragas]{PiCellPropData}}, and \code{\link[Ragas]{PiParentMetaData}}. When multiple analyses of the same type have been performed,
A \code{\link[Ragas]{PiDataList}} object will be created, which assembles several \code{PiData} objects of the same analysis/data type into one list.
A \code{\link[Ragas]{Pi}} object consists of different types of \code{PiDataList}.
}
\seealso{
Other related classes: \code{\link[Ragas]{PiExpFreqData}}, \code{\link[Ragas]{PiMarkerData}},
\code{\link[Ragas]{PiDSData}}, \code{\link[Ragas]{PiCellPropData}}, \code{\link[Ragas]{PiParentMetaData}}, and
\code{\link[Ragas]{PiDataList}}.
}
