% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{PiDataList}
\alias{PiDataList}
\alias{CheckPiData.PiDataList}
\alias{print.PiDataList}
\alias{[[.PiDataList}
\alias{AddPiData.PiDataList}
\title{PiDataList Class}
\usage{
PiDataList(...)

\method{CheckPiData}{PiDataList}(x, seurat.obj, ...)

\method{print}{PiDataList}(x, ...)

\method{[[}{PiDataList}(x, i)

\method{AddPiData}{PiDataList}(x, value)
}
\arguments{
\item{...}{\code{\link[Ragas]{PiData}} objects or additional parameters to \code{PiDataList} methods}

\item{x}{A PiDataList object}

\item{seurat.obj}{A Seurat object}

\item{i}{Index for the list element to be extracted}

\item{value}{A \code{PiData} object}
}
\description{
PiDataList is a simple list of \code{\link[Ragas]{PiData}} objects
}
