% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{PiDSData}
\alias{PiDSData}
\alias{CheckPiData.PiDSData}
\alias{print.PiDSData}
\alias{[[.PiDSData}
\title{PiDSData Class}
\usage{
PiDSData(
  data,
  ident,
  group,
  sample,
  block = NULL,
  method,
  group.1,
  group.2,
  contrast
)

\method{CheckPiData}{PiDSData}(x, seurat.obj)

\method{print}{PiDSData}(x, ...)

\method{[[}{PiDSData}(x, i)
}
\arguments{
\item{data}{Output from \code{\link[Ragas]{RunPseudobulkAnalysis}}}

\item{ident}{Name for the identity data}

\item{group}{Name of the group from Seurat metadata for differential state analysis}

\item{sample}{Name of the metadata column to identify samples}

\item{block}{Blocking variable, such as Batch}

\item{method}{A character string describing the analysis method}

\item{group.1}{Group name for case}

\item{group.2}{Group name for control}

\item{contrast}{A character string summarizing the comparison}

\item{x}{A PiDSData object}

\item{seurat.obj}{A Seurat object}

\item{...}{Additional arguments}

\item{i}{Index for the list element to be extracted}
}
\description{
The PiDSData inherts \code{\link[Ragas]{PiData}} class, which stores differential state analysis results from \code{\link[Ragas]{RunPseudobulkAnalysis}}.
}
\examples{
\dontrun{
## Run pseudobulk analysis and construct a PiDSData object
gp1 <- "cSLE"; gp2 <- "cHD";
res <- RunPseudobulkAnalysis(csle.bcell.small,
                             ident.var = "seurat_clusters",
                             group.var = "Groups",
                             sample.var = "Names",
                             group.1 = gp1,
                             group.2 = gp2,
                             pbDS.method = "edgeR")
ds <- PiDSData(data = res,
               ident = "seurat_clusters",
               method = "edgeR",
               group = "Groups",
               sample = "Names",
               group.1 = gp1,
               group.2 = gp2,
               contrast = paste(gp1, "-",  gp2, sep = ""),
               block = NULL)

## Print object
ds

## Check object validity
CheckPiData(ds, seurat.obj = csle.bcell.small)

## Add PiDSData object to an existing Pi object
my.pi <- CreatePostIntegrationObject(csle.bcell.small) ## a minimum Pi object
my.pi <- AddPiData(my.pi, ds)
}

}
