% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{PiExpFreqData}
\alias{PiExpFreqData}
\alias{CheckPiData.PiExpFreqData}
\alias{print.PiExpFreqData}
\alias{[[.PiExpFreqData}
\title{PiExpFreqData Class}
\usage{
PiExpFreqData(data, ident, group = NULL, cutoff = 0)

\method{CheckPiData}{PiExpFreqData}(x, seurat.obj)

\method{print}{PiExpFreqData}(x, ...)

\method{[[}{PiExpFreqData}(x, i)
}
\arguments{
\item{data}{A list containing per-gene expression frequencies for each identity class}

\item{ident}{Name for the identity data}

\item{group}{Factor name to group cells by}

\item{cutoff}{Log-expression cutoff (default: 0)}

\item{x}{A PiExpFreqData object}

\item{seurat.obj}{A Seurat object}

\item{...}{Additional arguments}

\item{i}{Index for the list element to be extracted}
}
\description{
The PiExpFreqData inherts \code{\link[Ragas]{PiData}} class, which is used to store per gene expression frequency data calculated by \code{\link[Ragas]{CalculateExpFreqs}}.
}
\examples{
\dontrun{
## Calculate gene expression frequency and construct a PiExpFreqData object
res <- CalculateExpFreqs(csle.bcell.small,
                         ident = "seurat_clusters",
                         group.by = NULL,
                         cutoff = 0)
ef <- PiExpFreqData(data = res,
                    ident = "seurat_clusters",
                    group = NULL,
                    cutoff = 0)

## Print object
ef

## Check object validity
CheckPiData(ef, seurat.obj = csle.bcell.small)

## Add PiExpFreqData object to an existing Pi object
my.pi <- CreatePostIntegrationObject(csle.bcell.small) ## a minimum Pi object
my.pi <- AddPiData(my.pi, ef)
}

}
