% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi.R
\name{PiParentMetaData}
\alias{PiParentMetaData}
\alias{CheckPiData.PiParentMetaData}
\alias{print.PiParentMetaData}
\alias{[[.PiParentMetaData}
\title{PiParentMetaData Class}
\usage{
PiParentMetaData(data, parent.name)

\method{CheckPiData}{PiParentMetaData}(x, seurat.obj)

\method{print}{PiParentMetaData}(x, ...)

\method{[[}{PiParentMetaData}(x, i)
}
\arguments{
\item{data}{A data.frame for metadata}

\item{parent.name}{Name of the parent object}

\item{x}{A PiParentMetaData object}

\item{seurat.obj}{A Seurat object}

\item{...}{Additional arguments}

\item{i}{Index for the list element to be extracted}
}
\description{
The PiParentMetaData inherts \code{\link[Ragas]{PiData}} class, which stores metadata of the parent Seurat object.
}
\examples{
\dontrun{
## Construct a PiParentMetaData object from the parent Seurat object csle.pbmc.small
pm <- PiParentMetaData(data = csle.pbmc.small[[]],
                       parent.name = "csle.pbmc.small")

## Print object
pm

## Check object validity
## Cells in Seurat object csle.bcell.small is a subset of cells from its PBMC parent
CheckPiData(pm, csle.bcell.small)

## Add PiParentMetaData object to an existing Pi object
my.pi <- CreatePostIntegrationObject(csle.bcell.small) ## a minimum Pi object
my.pi <- AddPiData(my.pi, pm)
}

}
