% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProportionPlot.R
\name{ProportionPlot}
\alias{ProportionPlot}
\title{Run Cell Proportion Analysis and Make Plots}
\usage{
ProportionPlot(
  object,
  var.group,
  method = "unpooled",
  plot.all = TRUE,
  colors = NULL,
  prop.by = "cluster",
  split.by = NULL,
  parent.meta = NULL,
  unpool.plot.type = "boxplot",
  var.unpool,
  n.col = 5,
  plot.p.sig = TRUE,
  plot.p.adj.sig = FALSE,
  title.text.size = 18,
  axis.title.size = 15,
  axis.text.size = 12,
  axis.text.angle = 90,
  axis.text.hjust = 0,
  axis.text.vjust = 0,
  legend.title.size = 15,
  legend.text.size = 12,
  point.size = 1,
  sig.size = 8,
  sig.bracket.color = "gray50",
  sig.color = "gray50",
  sig.bracket.width = 0.5,
  random.col.seed = 1
)
}
\arguments{
\item{object}{A Seurat object}

\item{var.group}{Name of the metadata column to group cells by}

\item{method}{Method for cell proportion analysis: either "unpooled" or "pooled"}

\item{plot.all}{Whether to add proportion for all cells. Only applicable to pooled method (default: TRUE)}

\item{colors}{Colors for the plotting group}

\item{prop.by}{Choose one of the two methods to calculate pooled cell proportion: by "cluster" or "group"}

\item{split.by}{Name of the metadata column to split the pooled proportion plot by (optional)}

\item{parent.meta}{Meta data inherited from the parent object}

\item{unpool.plot.type}{Method for the unpooled plot: either boxplot or barplot}

\item{var.unpool}{Name of the secondary grouping variable from the metadata column, such as sample name}

\item{n.col}{Number of columns per plot (default: 5)}

\item{plot.p.sig}{Whether to highlight tests with significant raw p-values}

\item{plot.p.adj.sig}{Whether to highlight tests with significant adjusted p-values}

\item{title.text.size}{Text size for plot title}

\item{axis.title.size}{Axis title size}

\item{axis.text.size}{Axis text size}

\item{axis.text.angle}{Rotation of axis text}

\item{axis.text.hjust}{Horizontal justification (in [0, 1]) for axis text}

\item{axis.text.vjust}{Vertical justification (in [0, 1]) for axis text}

\item{legend.title.size}{Legend title size}

\item{legend.text.size}{Legend text size}

\item{point.size}{Point size for box plot}

\item{sig.size}{Asterisk size for significant tests}

\item{sig.bracket.color}{Bracket color for significant tests}

\item{sig.color}{Asterisk color for significant tests}

\item{sig.bracket.width}{Bracket width for significant tests}

\item{random.col.seed}{Set seed to control colors}
}
\description{
An internal function to run pooled or unpooled proportion analysis and make bar plot or box plot.
Use its wrapper function \code{\link[Ragas]{RunProportionPlot}} for plotting.
}
\keyword{internal}
