% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAnnotatedDotPlot.R
\name{RunAnnotatedDotPlot}
\alias{RunAnnotatedDotPlot}
\title{Run Annotated DotPlot for Seurat or post-integration (Pi) object}
\usage{
RunAnnotatedDotPlot(
  object,
  assay = NULL,
  group.by = "seurat_clusters",
  annotations = NULL,
  annotation.cols = NULL,
  features = NULL,
  cols = c("lightgrey", "blue"),
  split.by = NULL,
  clust.row = TRUE,
  clust.column = FALSE,
  column.fontsize = 12,
  row.fontsize = 12,
  legend.label.fontsize = 13,
  legend.title.fontsize = 15,
  random.col.seed = 42,
  ...
)
}
\arguments{
\item{object}{A Seurat or \code{\link[Ragas]{Pi}} object}

\item{assay}{Name of the assay to use (default: the active assay as indicated by  \code{\link[Seurat]{DefaultAssay}})}

\item{group.by}{A factor that contains identity information to group the cells by (default: seurat_clusters)}

\item{annotations}{User defined annotations in one of the following two formats:
\itemize{
\item{list} : a named list object. Each list element is a character vector of a group of features; the name of the list element will be extracted to represent the feature group
\item{data.frame} : a data.frame with two columns: "features" and "annotation"
}}

\item{annotation.cols}{Colors for feature annotations. Needs to be a character vector (preferably a named vector with its names matching the "annotations" argument. See example below) of the same length as annotation feature groups; will be automatically assigned with random colors if set to NULL}

\item{features}{a list of features/genes to plot without annotations. Will be ignored if argument annotations is set}

\item{cols}{Colors to plot. Same as the "cols" argument of \code{\link[Seurat]{DotPlot}} (default: c("lightgrey", "blue"))}

\item{split.by}{Factor to split the groups by. Same as the "split.by" argument of \code{\link[Seurat]{DotPlot}}}

\item{clust.row}{Whether to cluster the rows/identities (default: TRUE)}

\item{clust.column}{Whether to cluster the columns/features (default: FALSE)}

\item{column.fontsize}{Size of column text (default: 12)}

\item{row.fontsize}{Size of row text (default: 12)}

\item{legend.label.fontsize}{Size of the legend labels (default: 13)}

\item{legend.title.fontsize}{Size of the legend title (default: 15)}

\item{random.col.seed}{Set seed to control colors}

\item{...}{Extra parameters passed to DotPlot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
A wrapper function for \code{\link[Ragas]{AnnotatedDotPlot}}
}
\examples{
\dontrun{
my.list <- list(Monocytes = c("CD14","FCGR3A"),
                B = "MS4A1",
                T = c("CCR7", "CD8A"),
                NK = "NKG7")
RunAnnotatedDotPlot(object = csle.pbmc.pi,
                    annotation.cols = c( 'Monocytes'= 'tomato', 'B' = 'seagreen', 'T' = 'steelblue', 'NK' = 'purple'),
                    annotations = my.list)
}

}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
