% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDimPlot.R
\name{RunDimPlot}
\alias{RunDimPlot}
\title{Run DimPlot}
\usage{
RunDimPlot(
  object,
  group.by = "seurat_clusters",
  reduction = "umap",
  label = TRUE,
  repel = TRUE,
  pt.size = 0.1,
  label.size = 3,
  random.col.seed = 42,
  cols = NULL,
  ggrepel.max.overlaps = Inf,
  ...
)
}
\arguments{
\item{object}{A Seurat or \code{\link[Ragas]{Pi}} object}

\item{group.by}{Name of the metadata column to group (color) cells by (default: seurat_clusters)}

\item{reduction}{Which dimension reduction to use (default: umap)}

\item{label}{Whether to show the label of clusters (default: TRUE)}

\item{repel}{Whether to repel labels (default: TRUE)}

\item{pt.size}{Plot point size (default: 0.1)}

\item{label.size}{Label size (default: 3)}

\item{random.col.seed}{Set seed to control colors}

\item{cols}{User-defined colors. When cols is set, random.col.seed will be ignored}

\item{ggrepel.max.overlaps}{Adjust overlap for ggrepel (default: Inf)}

\item{...}{Extra parameters passed to DimPlot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Run DimPlot for Seurat or post-integration (Pi) object
}
\details{
Wrapper functiuon for \code{\link[Seurat]{DimPlot}} that is enhanced to: (1) uses \code{\link[randomcoloR]{distinctColorPalette}} to generate random colors;
(2) allows adjustment of overlap among identity labels. By default, column \code{seurat_clusters} from the metadata
will be by assigned as \code{group.by}.
}
\examples{
\dontrun{
## default, use "seurat_clusters" to group cells
RunDimPlot(object = csle.pbmc.pi)

## use subcluster identities and re-projected UMAP
RunDimPlot(object = csle.pbmc.pi, group.by = "subcluster_idents", reduction = "rp", label.size = 5, pt.size = 1)
}
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Ron Ammar (2019). randomcoloR: Generate Attractive Random Colors. R package version 1.1.0.1.https://CRAN.R-project.org/package=randomcoloR
}
