% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDSDotPlot.R
\name{RunDSDotPlot}
\alias{RunDSDotPlot}
\title{Run DSDotPlot}
\usage{
RunDSDotPlot(
  object,
  features = NULL,
  exp.freq.key,
  ds.key,
  p.filter = 0.05,
  FC.filter = 2,
  exp.freq.filter = 0.1,
  filter.by.median.diff = TRUE,
  median.diff.filter = 0,
  to.adjust = FALSE,
  top.n = Inf,
  direction = "both",
  to.highlight = TRUE,
  clust.row = TRUE,
  clust.column = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{\link[Ragas]{Pi}} object}

\item{features}{A list of features/genes to plot}

\item{exp.freq.key}{A unique identifier to retrieve data from the \code{exp.freq} object. See  \code{\link[Ragas]{CalculateExpFreqs}} for more details}

\item{ds.key}{A unique identifier to retrieve data from the \code{ds} object. See  \code{\link[Ragas]{RunPseudobulkAnalysis}} for more details}

\item{p.filter}{P-Value cutoff for gene filtering (default: 0.05)}

\item{FC.filter}{Fold change cutoff for gene filtering (default: 2)}

\item{exp.freq.filter}{Gene expression frequency filtering cutoff (default: 0.1)}

\item{filter.by.median.diff}{Whether to perform additional filtering based on median CPM differences between groups. This will remove differential genes due to outliers.}

\item{median.diff.filter}{Median difference filtering cutoff (default: 0)}

\item{to.adjust}{Whether to perform adjustment for multiple testing (default: FALSE)}

\item{top.n}{Number of top genes to plot (default: Inf, i.e., all genes passing filtering criteria will be plotted)}

\item{direction}{Whether to plot "up" or "down" regulated genes, or "both" (default: "both")}

\item{to.highlight}{Whether to highlight feature/identity pairs that passed the filtering criteria (default: TRUE)}

\item{clust.row}{Whether to cluster rows (default: TRUE)}

\item{clust.column}{Whether to cluster columns (default: FALSE)}

\item{verbose}{Verbosity (default: TRUE)}

\item{...}{Extra parameters passed to \code{\link[Ragas]{DSDotPlot}} to change plot font and color, etc.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
A wrapper function for \code{\link[Ragas]{DSDotPlot}}
}
\details{
The input Pi object must have analyzed results in the \code{ds} and \code{exp.freq} fields, which can be accomplished by running \code{\link[Ragas]{RunPseudobulkAnalysis}}
and \code{\link[Ragas]{CalculateExpFreqs}}, respectively.

Gene filtering will be performed based on significance (i.e., \code{p.filter} and \code{FC.filter}) and expression frequency (\code{exp.freq.filter}).
The purpose of significance filtering is twofold: (1) only features that passed the filtering criteria in at least one of the identity groups (e.g., cell clusters,
in most cases) will be included in the plot; (2) significant features that passed the filtering criteria will be highlighted on their corresponding clusters.
The \code{FC.filter} argument expects un-logged fold-change and a \code{FC.filter} of 2 can mean a two-fold expression change for both up or down-regulation.
If \code{top.n} is set to a number other than Inf, only the top genes will be plotted.

If the argument \code{features} is set, all valid features (i.e., exist in the Seurat object) will be plotted in the figure regardless of their significance. However, only features
passed the filtering criteria will be highlighted.
}
\examples{
\dontrun{
RunDSDotPlot(object = csle.bcell.pi,
             exp.freq.key = "ExpFreq|cluster.annotation|cutoff=0",
             ds.key = "DS|cluster.annotation|edgeR|group=Groups;sample=Names;gp1=cSLE;gp2=cHD;contrast=cSLE-cHD",
             p.filter = 0.05)

}
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
