% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunExpressionPlot.R
\name{RunExpressionPlot}
\alias{RunExpressionPlot}
\title{Run Expression Plot}
\usage{
RunExpressionPlot(
  object,
  feature = NULL,
  assay = "RNA",
  group.by = NULL,
  split.by = NULL,
  ident = "seurat_clusters",
  ds.key = NULL,
  plot.ds.p.value = FALSE,
  plot.ds.adjp.value = FALSE,
  colors = NULL,
  random.col.seed = 1,
  point.size = 1,
  axis.text.size = 12,
  axis.text.angle = 90,
  axis.text.hjust = 0,
  axis.text.vjust = 0,
  axis.title.size = 15,
  legend.text.size = 12,
  legend.title.size = 15,
  sig.size = 8,
  sig.bracket.color = "gray50",
  sig.color = "gray50",
  sig.bracket.width = 0.5
)
}
\arguments{
\item{object}{A Seurat object or a \code{\link[Ragas]{Pi}} object}

\item{feature}{Gene name}

\item{assay}{Name of the assay to use (default: "RNA")}

\item{group.by}{Name of metadata column to group cells by}

\item{split.by}{Name of the metadata column to split expression by (optional)}

\item{ident}{Cell identity variable (default: seurat_clusters)}

\item{ds.key}{A named list with unique identifier to retrieve data from the \code{ds} object. List names must be of format group1_group2 where group1 and group2 are groups in the split.by argument. See  \code{\link[Ragas]{RunPseudobulkAnalysis}} for more details to create ds object.}

\item{plot.ds.p.value}{TRUE/FALSE if p-value from differential testing should be plotted (default: FALSE)}

\item{plot.ds.adjp.value}{TRUE/FALSE if adjusted p-value from differential testing should be plotted (default: FALSE)}

\item{colors}{Colors for plotting. Number of colors must be the equal to number of idents or the number of groups if split.by is used (optional)}

\item{random.col.seed}{Set seed to control colors}

\item{point.size}{Point size for box plot}

\item{axis.text.size, legend.text.size}{Text size for axis or legend (default: 12)}

\item{axis.title.size, legend.title.size}{Title size for axis or legend (default: 15)}

\item{sig.size}{Asterisk size for significant tests}

\item{sig.bracket.color}{Bracket color for significant tests}

\item{sig.color}{Asterisk color for significant tests}

\item{sig.bracket.width}{Bracket width for significant tests}
}
\value{
A \code{\link[ggplot2]{ggplot}} object will be returned
}
\description{
A wrapper function for \code{\link[Ragas]{ExpressionPlot}}
}
\details{
Plot expression levels of gene per identity class. The average expression level is calculated by the \code{group.by} argument. If \code{split.by} is set, plot will be
further separated by data defined by the \code{split.by} argument.

To provide user-defined colors, a vector of colors is required. Number of colors must be equal to number of idents or
number of groups if \code{split.by} is used.
}
\examples{
\dontrun{
# Seurat input and user defined colors
library(scales)
cols <-  hue_pal()(length(levels(csle.pbmc.small$Groups)))
RunExpressionPlot(object = csle.pbmc.small,
                  feature = "ISG15",
                  ident = "cluster.annotation",
                  group.by = "Names",
                  split.by = "Groups",
                  colors = cols)

# Pi input without ds p-value
RunExpressionPlot(object = csle.pbmc.pi,
                  feature = "ISG15",
                  ident = "subcluster_idents",
                  group.by = "Names",
                  split.by = "Groups")

# Pi input with ds p-value
RunExpressionPlot(object = csle.pbmc.pi,
                  feature = "ISG15",
                  ident = "subcluster_idents",
                  group.by = "Names",
                  ds.key = list(cSLE_cHD = "DS|subcluster_idents|edgeR|group=Groups;sample=Names;gp1=cSLE;gp2=cHD;contrast=cSLE-cHD"),
                  plot.ds.p.value = TRUE,
                  plot.ds.adjp.value = FALSE,
                  split.by = "Groups")
}
}
