% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFindAllMarkers.R
\name{RunFindAllMarkers}
\alias{RunFindAllMarkers}
\title{Run FindAllMarkers from Seurat}
\usage{
RunFindAllMarkers(
  object,
  ident = "seurat_clusters",
  test.use = "wilcox",
  latent.vars = NULL,
  future.strategy = "multisession",
  future.workers = 1,
  future.globals.maxSize.MB = 1000,
  ...
)
}
\arguments{
\item{object}{A Seurat or \code{\link[Ragas]{Pi}} object}

\item{ident}{Name of the metadata column used as cell identity (default: seurat_clusters)}

\item{test.use}{Which test to use. See details in \code{\link[Seurat]{FindAllMarkers}} (default: wilcox)}

\item{latent.vars}{Latent variables to test. See details in \code{\link[Seurat]{FindAllMarkers}}}

\item{future.strategy}{Parallelization strategies. See \code{\link[future]{plan}} for more details (default: "multisession)}

\item{future.workers}{Number of parallelizations (default: 1)}

\item{future.globals.maxSize.MB}{Maximum allowed total size (in Megabyte) of global variables. See \code{\link[future]{future.options}} (default: 1000)}

\item{...}{Additional parameters passed to \code{\link[Seurat]{FindAllMarkers}}}
}
\value{
If \code{object} is a Seurat object, a matrix containing markers and their associated statistics will be returned
(see \code{\link[Seurat]{FindMarkers}} for more details); If \code{object} is a Pi object, a \code{\link[Ragas]{PiMarkerData}} object containing
the marker result matrix will be created and attached to the \code{markers} field.
}
\description{
This is a wrapper function to run \code{\link[Seurat]{FindAllMarkers}} function on Seurat or Pi object
}
\examples{
\dontrun{
my.pi <- CreatePostIntegrationObject(object = csle.pbmc.small) ## a minimum Pi object
my.pi <- RunFindAllMarkers(object = my.pi)
}

}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
