% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMatrixPlot.R
\name{RunMatrixPlot}
\alias{RunMatrixPlot}
\title{Run Matrix Plot}
\usage{
RunMatrixPlot(
  object,
  markers.key,
  assay = NULL,
  alt.assay.name = "RNAalt",
  top.n = 5,
  up.genes = TRUE,
  heatmap.cols = NULL,
  min.exp = -1.5,
  max.exp = 1.5,
  column.fontsize = 6,
  row.fontsize = 8,
  column.fontface = "plain",
  row.fontface = "plain",
  column.anno.cols = NULL,
  column.anno.name.fontsize = 10,
  column.anno.name.rot = 0,
  column.anno.name.just = "left",
  legend.label.fontsize = 10,
  legend.title.fontsize = 10,
  heatmap.width = 20,
  heatmap.height = 8
)
}
\arguments{
\item{object}{A \code{\link[Ragas]{Pi}} object}

\item{markers.key}{A unique identifier for \code{\link[Ragas]{RunFindAllMarkers}} to pull results from the \code{markers} field}

\item{assay}{Seurat assay name}

\item{alt.assay.name}{Name of the alternative assay to store RC normalized data}

\item{top.n}{Number of top markers to plot (per identity group)}

\item{up.genes}{Whether to plot the up-regulated genes (default: TRUE)}

\item{heatmap.cols}{Colors to plot the heatmap (optional). A character vector with colors for low, medium and high expression.}

\item{min.exp}{The minimum expression value to plot (default: -1.5)}

\item{max.exp}{The maximum expression value to plot (default: 1.5)}

\item{column.fontsize}{Font size for column texts (default: 6)}

\item{row.fontsize}{Font size for row texts (default: 8)}

\item{column.fontface}{Fontface for column labels, either "plain", "bold", "italic", "oblique" or "bold.italic" (default: "plain")}

\item{row.fontface}{Fontface for column labels, either "plain", "bold", "italic", "oblique" or "bold.italic" (default: "plain")}

\item{column.anno.cols}{Named character vector with colors for column annotations (default: NULL). Names should match Idents of object.}

\item{column.anno.name.fontsize}{Font size for column annotations, such as clusters (default: 10)}

\item{column.anno.name.rot}{Rotation of column annotations (default: 0)}

\item{column.anno.name.just}{Alignment of column annotations (default: "left")}

\item{legend.label.fontsize}{Font size for legend label (default: 10)}

\item{legend.title.fontsize}{Font size for legend title (default: 10)}

\item{heatmap.width}{Width of the heatmap (default: 20)}

\item{heatmap.height}{Height of the heatmap (default: 8)}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
\description{
A wrapper function for \code{\link[Ragas]{MatrixPlot}}
}
\details{
This function takes pre-calculated results from the \code{markers} field of the Pi object and plots heatmap of
top genes for each identity group (e.g., cluster). Running \code{\link[Ragas]{RunFindAllMarkers}} is always required prior to run this function,
which will store new  \code{\link[Seurat]{FindAllMarkers}} results in the \code{markers} object with a unique identifier.
This unique identifier needs to be passed to the \code{markers.key} argument so the function can pull relevant differential
expression data for plotting.

A new assay with assay name determined by argument \code{alt.assay.name} will be created, followed by
RC normalization and scaling. When scaling, no variables are regressed out so the scaled expression will be more consistent with that from
\code{\link[Seurat]{FeaturePlot}}. The new assay only contains \code{top.n} genes from each identity group (e.g., clusters, in most cases) to
reduce computational cost.
}
\examples{
\dontrun{
## Run matrix plot for pre-calculated markers in the example data
RunMatrixPlot(csle.pbmc.pi,
              markers.key = "Markers|subcluster_idents|AllMarkers|test.use=wilcox",
              column.anno.name.rot = 45)
}

}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
