% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProportionPlot.R
\name{RunProportionPlot}
\alias{RunProportionPlot}
\title{Run ProportionPlot}
\usage{
RunProportionPlot(
  object,
  group.by,
  ident = "seurat_clusters",
  method = c("unpooled", "pooled"),
  group.colors = NULL,
  use.parent.as.ref = FALSE,
  parent.meta.data.key = NULL,
  return.value = c("stats", "ggplot"),
  pooled.prop.by = c("cluster", "group"),
  pooled.split.by = NULL,
  pooled.add.all = TRUE,
  unpool.by = NULL,
  unpool.plot.type = c("boxplot", "barplot"),
  unpool.ncol = 5,
  unpool.plot.p.sig = TRUE,
  unpool.plot.p.adj.sig = FALSE,
  unpool.pt.size = 1,
  unpool.sig.size = 6,
  unpool.sig.bracket.color = "black",
  unpool.sig.color = "black",
  unpool.sig.bracket.width = 0.5,
  title.text.size = 12,
  axis.title.size = 12,
  axis.text.size = 12,
  legend.title.size = 12,
  legend.text.size = 12,
  axis.text.angle = 90,
  axis.text.hjust = 1,
  axis.text.vjust = 1,
  random.col.seed = 1
)
}
\arguments{
\item{object}{A Seurat or \code{\link[Ragas]{Pi}} object}

\item{group.by}{Name of the metadata column to group cells by}

\item{ident}{Cell identity variable (default: seurat_clusters)}

\item{method}{Method for cell proportion analysis: either "unpooled" or "pooled" (default: "unpooled")}

\item{group.colors}{Colors for the plotting group}

\item{use.parent.as.ref}{Whether to use cell counts from the parent object to calculate cell proportions (default: FALSE)}

\item{parent.meta.data.key}{A unique identifier to retrieve parent metadata}

\item{return.value}{Type of the return value (default: "stats")}

\item{pooled.prop.by}{Choose one of the two methods to calculate pooled cell proportion: by "cluster" or "group" (default: "cluster")}

\item{pooled.split.by}{Name of the metadata column to split the pooled proportion plot by (optional)}

\item{pooled.add.all}{Whether to add proportion for all cells (default: TRUE)}

\item{unpool.by}{Name of the secondary grouping variable from the metadata column, such as sample name}

\item{unpool.plot.type}{Method for the unpooled plot: either boxplot or barplot (default: "boxplot")}

\item{unpool.ncol}{Number of columns per plot (default: 5)}

\item{unpool.plot.p.sig}{Whether to highlight tests with significant raw p-values (default: TRUE)}

\item{unpool.plot.p.adj.sig}{Whether to highlight tests with significant adjusted p-values (default: FALSE)}

\item{unpool.pt.size}{Point size for box plot (default: 1)}

\item{unpool.sig.size, unpool.sig.bracket.color, unpool.sig.color, unpool.sig.bracket.width}{Additional aesthetic parameters that control highlighting of significant tests}

\item{title.text.size, axis.title.size, axis.text.size, legend.title.size, legend.text.size}{Text size for plot title, axis, or legend (default: 12)}

\item{axis.text.angle}{Rotation for axis text (default: 90)}

\item{axis.text.hjust, axis.text.vjust}{Horizontal/vertical justification (in [0, 1]) for axis text (default: 1)}

\item{random.col.seed}{Set seed to control colors}
}
\value{
When \code{return.value} is "ggplot", a ggplot object will be returned; when \code{return.value} is "stats" and \code{object} is a Seurat object,
a data summary list is returned. If \code{method} is "pooled", the data summary list only contains cell proportions used for the pooled plot; If \code{method} is "unpooled",
the summary list encloses an additional data.frame summarizing test statistics for cell proportions,
which contains the following 11 columns for all pairwise cell proportion comparisons:
\itemize{
\item{Cluster} : Name of cell cluster/identity
\item{group1} : Name of group 1
\item{group2} : Name of group 2
\item{estimate1} : Estimated proportion of group1
\item{estimate2} : Estimated proportion of group2
\item{estimate} : Estimated proportion difference, i.e., estimate1 - estimate2
\item{statistic} : Test statistic
\item{p} : P-value
\item{p.adj} : Adjusted P-value (method: "BH")
\item{p.signif} : A Boolean variable indicating whether the a test is significant based on P-value
\item{p.adj.signif} : A Boolean variable indicating whether the a test is significant based on adjusted P-value
}
When \code{return.value} is "stats" and \code{object} is a Pi object, an updated Pi object will be returned: a \code{\link[Ragas]{PiCellPropData}}
object that encloses the summary list will be created and added to the \code{cell.prop} field.
}
\description{
A wrapper function for \code{\link[Ragas]{ProportionPlot}}
}
\details{
This function supports two methods for single-cell proportion analysis, namely "pooled" and "unpooled". The pooled method
will pool all cells across all samples and calculate cell proportion per cluster per group (as determined by the argument \code{group.by}).
Only bar plot is applicable for the pooled method.

For the pooled method, proportions can be calculated by two ways, either by cluster or by group, given by the \code{pooled.prop.by} argument.
If \code{pooled.prop.by} is set to "cluster", proportions of cells per cluster per group will be "normalized" by the total number of cells in each cluster.
In this case, for a given cluster, the proportions of cells from all groups will sum up to 1. This will lead to stacked bar plot for each cluster.
If \code{pooled.prop.by} is set to "group", proportion of cells for per cluster per group will be "normalized" by the total number of cells in each group, combining all cell clusters.
This will lead to grouped bar plots.

The "unpooled" method is suitable for multi-sample, multi-group study design, where an additional grouping argument \code{unpool.by} will decide how to split cells
into biological replicates. A typical choice for \code{unpool.by} is the metadata column that contains the sample/replicate IDs/names.
Two plotting methods are supported under the unpooled mode: box plot and bar plot. For box plot, T test of cell proportion differences will be performed between all groups
using the \code{\link[rstatix]{t_test}} function and significant proportional differences will be highlighted between corresponding groups; for bar plot,
standard errors will be plotted.

If \code{use.parent.as.ref} is set to TRUE, cell numbers from the parent object will be use as reference (i.e., denominator) for cell proportion calculation.
This will give users the flexibility to investigate cell proportions for any cluster relative to a given choice of "total population". For example, when
comparing cell proportions of memory CD4 T cells, if \code{use.parent.as.ref} is FALSE, then the derived cell proportion per cluster means percentage of
cells within the total memory CD4 cells; if \code{use.parent.as.ref} is TRUE and a PBMC object is assigned as \code{parent.object} when running
\code{\link[Ragas]{CreatePostIntegrationObject}}, then the calculated percentage means percentage of cell in total PBMC. The \code{use.parent.as.ref}
argument only affect unpooled methods and pooled method with \code{pooled.prop.by} set to "group".
}
\examples{
\dontrun{
## Create a Pi object for b cells and assign parent object to the PBMC
my.pi <- CreatePostIntegrationObject(object = csle.bcell.small,
                                     parent.object = csle.pbmc.small)

## pooled, proportion by cluster
my.pi <- RunProportionPlot(my.pi,
                           ident = "cluster.annotation",
                           group.by = "Groups",
                           method = "pooled")

## pooled and split, proportion by group
my.pi <- RunProportionPlot(my.pi,
                           ident = "cluster.annotation",
                           group.by = "Groups",
                           method = "pooled",
                           pooled.prop.by = "group",
                           pooled.split.by = "Ethnicity")

## unpooled, boxplot, use parent (pbmc) to normalize cell counts
my.pi <- RunProportionPlot(my.pi,
                           ident = "cluster.annotation",
                           group.by = "Groups",
                           method = "unpooled",
                           unpool.by = "Names",
                           use.parent.as.ref = TRUE)

## unpooled, barplot
my.pi <- RunProportionPlot(my.pi,
                           ident = "cluster.annotation",
                           group.by = "Groups",
                           method = "unpooled",
                           unpool.by = "Names",
                           unpool.plot.type = "barplot")

## Input is a Seurat object, output is a ggplot
RunProportionPlot(csle.bcell.small,
                  ident = "cluster.annotation",
                  group.by = "Groups",
                  method = "unpooled",
                  unpool.by = "Names",
                  return.value = "ggplot")
}

}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
