% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPseudobulkAnalysis.R
\name{RunPseudobulkAnalysis}
\alias{RunPseudobulkAnalysis}
\title{Run Pseudobulk Analysis}
\usage{
RunPseudobulkAnalysis(
  object,
  ident.var,
  group.var,
  sample.var,
  blocking.var = NULL,
  group.1,
  group.2,
  pbDS.method = c("edgeR", "DESeq2", "limma-trend", "limma-voom"),
  pbDS.verbose = TRUE,
  pbDS.min.cell = 1,
  pbDS.filter = "none",
  resDS.bind = "row",
  resDS.frq = TRUE,
  resDS.cpm = TRUE,
  resDS.digits = Inf,
  resDS.sep = "__",
  calcExprFreqs.assay = "counts",
  calcExprFreqs.th = 0,
  auto.make.names = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat or \code{\link[Ragas]{Pi}} object}

\item{ident.var}{Cell identity variable}

\item{group.var}{Group variable, used to test differential state}

\item{sample.var}{Sample variable}

\item{blocking.var}{Additional blocking variable used for batch correction (default: NULL)}

\item{group.1}{A character string for the case group, must be consistent with the levels of \code{group.var}}

\item{group.2}{A character string for the ctrl group, must be consistent with the levels of \code{group.var}}

\item{pbDS.method}{Name of the method for pseudobulk analysis (default: edgeR)}

\item{pbDS.verbose}{Verbosity of the \code{\link[muscat]{pbDS}} function}

\item{pbDS.min.cell}{A numeric number that specifies the minimum number of cells required in a given cluster-sample.
See \code{\link[muscat]{pbDS}} for details (default 1)}

\item{pbDS.filter}{Whether to filter on genes and/or samples. See \code{\link[muscat]{pbDS}} for details (default: "none")}

\item{resDS.bind, resDS.frq, resDS.cpm, resDS.digits, resDS.sep}{parameters for \code{\link[muscat]{resDS}}}

\item{calcExprFreqs.assay, calcExprFreqs.th}{parameters for \code{\link[muscat]{calcExprFreqs}}}

\item{auto.make.names}{Whether to automatically check and make syntactically valid names (default: TRUE)}

\item{verbose}{Verbosity (default: TRUE)}

\item{...}{Extra parameters passed to \code{\link[muscat]{pbDS}}}
}
\value{
When \code{object} is a Seurat object, a list containing the output of \code{\link[muscat]{pbDS}} and  \code{\link[muscat]{resDS}} will be returned;
If \code{object} is a Pi object (recommended), an updated Pi object will be returned:  a \code{\link[Ragas]{PiDSData}} object that encloses pseudobulk analysis results
will be created and added to the \code{ds} field.
}
\description{
Run differential state analysis with the \code{muscat} package on
pseudobulk counts for multi-sample, multi-group study design
}
\examples{
\dontrun{
my.pi <- CreatePostIntegrationObject(object = csle.pbmc.small) ## a minimum Pi object
my.pi <- RunPseudobulkAnalysis(object = my.pi,
                               ident.var = "cluster.annotation",
                               group.var = "Groups",
                               sample.var = "Names",
                               group.1 = "cSLE",
                               group.2 = "cHD")
}
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048

Helena L. Crowell, Pierre-Luc Germain, Charlotte Soneson, Anthony Sonrel and Mark D. Robinson (2020). muscat: Multi-sample
multi-group scRNA-seq data analysis tools. R package version 1.4.0. https://github.com/HelenaLC/muscat

Ritchie, M.E., Phipson, B., Wu, D., Hu, Y., Law, C.W., Shi, W., and Smyth, G.K. (2015). limma powers differential expression analyses
for RNA-sequencing and microarray studies. Nucleic Acids Research 43(7), e47.
}
