% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSummarizedHeatmap.R
\name{RunSummarizedHeatmap}
\alias{RunSummarizedHeatmap}
\title{Run Summarized Heatmap for Seurat or post-integration (Pi) object}
\usage{
RunSummarizedHeatmap(
  object,
  features,
  assay = NULL,
  ident = "seurat_clusters",
  split.by = NULL,
  additional.metadata = NULL,
  heatmap.cols = c("cadetblue", "black", "yellow"),
  clust.row = TRUE,
  clust.column = TRUE,
  row.annotation.cols = NULL,
  column.annotation.cols = NULL,
  show.row.names = TRUE,
  show.column.names = TRUE,
  min.exp = -2,
  max.exp = 2,
  column.names.rotation = 90,
  row.fontsize = 10,
  column.fontsize = 8,
  row.fontface = "plain",
  column.fontface = "plain",
  legend.label.fontsize = 10,
  legend.title.fontsize = 10,
  annotation.name.fontsize = 10,
  heatmap.width = 12,
  heatmap.height = 10,
  random.col.seed = 1,
  ...
)
}
\arguments{
\item{object}{A Seurat or \code{\link[Ragas]{Pi}} object}

\item{features}{User defined feature list in one of the following two formats:
\itemize{
\item{list} : a named list object. Each list element is a character vector of a group of features; the name of the list element will be extracted to represent the feature group
\item{vector} : a character vector of features; no row annotation is applicable so leave \code{row.annotation.cols} to NULL
}}

\item{assay}{Name of the assay to use (default: the active assay as indicated by  \code{\link[Seurat]{DefaultAssay}})}

\item{ident}{Name of the metadata column to be used as cell identity (default: seurat_clusters)}

\item{split.by}{Factor to split the columns by}

\item{additional.metadata}{Names of additional metadata used for column annotation}

\item{heatmap.cols}{Colors for filling the heatmap (default: c("cadetblue","black","yellow"))}

\item{clust.row, clust.column}{Whether to cluster the rows/columns (default: TRUE)}

\item{row.annotation.cols}{A named list containing colors for column annotations; If set to NULL, \code{\link[randomcoloR]{distinctColorPalette}} will be used to generate random colors}

\item{column.annotation.cols}{A character vector containing colors for row annotations; If set to NULL, \code{\link[randomcoloR]{distinctColorPalette}} will be used to generate random colors}

\item{show.row.names, show.column.names}{Whether to show row/column names (default: TRUE)}

\item{min.exp, max.exp}{Minimum/maximum scaled expression value to plot (default: -2/2)}

\item{column.names.rotation}{Rotation for column names (default: 90)}

\item{row.fontsize, column.fontsize}{Row/column font size (default: 10/8)}

\item{row.fontface, column.fontface}{Fontface for row/column labels, either "plain", "bold", "italic", "oblique" or "bold.italic" (default: "plain")}

\item{legend.label.fontsize, legend.title.fontsize, annotation.name.fontsize}{Font sizes for legend label/title and annotation name (default: 10)}

\item{heatmap.width, heatmap.height}{Width and height of heatmap (default: 12/10)}

\item{random.col.seed}{Set seed to control row and column annotation colors (default: 1)}

\item{...}{Extra parameters passed to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
\description{
A wrapper function for \code{\link[Ragas]{SummarizedHeatmap}}
}
\details{
Plot heatmap summarizing expression levels per identity class. If \code{split.by} is set, the columns of the heatmap will be
further separated by data defined by the \code{split.by} argument. Additional metadata columns are also allowed to be included for column
annotation, provided that their levels are consistent with those of the \code{split.by} argument in a one/multiple (data to split) to one
(additional metadata) relationship. For example, if metadata column "patient" is used to split the columns of the heatmap, additional metadata like "age",
"gender", or "disease group" are allowed since one patient typically has unique values for age, gender, and the disease group; on the other hand,  if metadata
"disease group" is used to split the columns, one can no longer add "age", "gender", or "patient" as additional metadata since one level of the disease group
(i.e., healthy) can correspond to multiple patients with difference values for age and gender.

To provide user-defined colors for column annotations, a named list of color vectors is required, with each of the list element corresponds to annotation data
passed by arguments "ident", "split.by", or "additional.metadata". The list element name for argument "ident" should always be "Cluster", and use the corresponding
metadata column names for annotations provided by "split.by" or "additional.metadata" (see example).
}
\examples{
\dontrun{
# Example 1: Seurat input and named list for features
features <- list("T cell" = c("CD8A", "CD8B"),
                 "ISG" = c("ISG15","ISG20","IFI44L"),
                 "Housekeeping" = c("ACTB"))

RunSummarizedHeatmap(object = csle.pbmc.small,
                     ident = "cluster.annotation",
                     features = features,
                     split.by = "Groups")

# Example 2: a more complex example with more metadata and user-defined column and row annotation color
## make a small example data set
my.bcell <- subset(csle.bcell.small, subset = Names \%in\% c("cHD1", "CHD2", "cHD3","cSLE1", "cSLE2", "cSLE3"))
my.bcell$Names <- droplevels(my.bcell$Names)

## user-defined column annotation color
library(randomcoloR)
cols.cluster <- distinctColorPalette(length(levels(my.bcell$cluster.annotation))); names(cols.cluster) <- levels(my.bcell$cluster.annotation)
cols.name <- distinctColorPalette(length(levels(my.bcell$Names))); names(cols.name) <- levels(my.bcell$Names)
cols.group <- distinctColorPalette(length(levels(my.bcell$Groups))); names(cols.group) <- levels(my.bcell$Groups)
cols.batch <- distinctColorPalette(length(levels(factor(my.bcell$Batch)))); names(cols.batch) <- levels(factor(my.bcell$Batch))
cols.column.anno <- list("Cluster" = cols.cluster,
                         "Names" = cols.name,
                         "Groups" = cols.group,
                         "Batch" = cols.batch)
## user-defined row/feature annotation color
features <- list("B cell" = c("TCL1A","ZEB2","CD27","ITGAX"),
                 "ISG" = c("ISG15","ISG20","IFI44L"),
                 "Housekeeping" = c("ACTB"))
cols.features <- distinctColorPalette(length(features))

RunSummarizedHeatmap(object = my.bcell,
                     ident = "cluster.annotation",
                     features = features,
                     split.by = "Names",
                     additional.metadata = c("Groups","Batch"),
                     column.annotation.cols = cols.column.anno,
                     row.annotation.cols = cols.features )

# Example 3: Pi input and vector input for features
features <- c("CD8A","CD8B","ISG15","ISG20","IFI44L","ACTB")
RunSummarizedHeatmap(object = csle.pbmc.pi,
                     ident = "subcluster_idents",
                     features = features,
                     split.by = "Groups")
}
}
\references{
Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
