% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackedVlnPlot.R
\name{StackedVlnPlot}
\alias{StackedVlnPlot}
\title{Stacked Violin Plot}
\usage{
StackedVlnPlot(
  object,
  features,
  feature.annotation = NULL,
  feature.annotation.cols = NULL,
  assay = "RNA",
  split.by = NULL,
  color.by = "features",
  clust.row = FALSE,
  clust.column = TRUE,
  add.points = FALSE,
  points.size = 0.1,
  column.fontsize = 12,
  row.fontsize = 8,
  column.fontface = "plain",
  row.fontface = "plain",
  features.fontface = "plain",
  row.title.fontsize = 15,
  legend.fontsize = 12,
  legend.title.fontsize = 15,
  features.fontsize = 12,
  column.names.rotation = 0,
  axis.text.hjust = 0.5,
  axis.text.vjust = 0.5,
  colors = NULL,
  random.annotation.col.seed = 1
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{Features to plot}

\item{feature.annotation}{Annotation for features}

\item{feature.annotation.cols}{Annotation color}

\item{assay}{Name of the assay to use}

\item{split.by}{Factor to split the columns by}

\item{color.by}{One of the five ways to color for the violin plot: "features", "clusters", "median.exp" (median expression), "mean.exp" (mean expression), "split.var" (the variable that splits the violin plot).
When \code{split.by} is set (not to NULL), \code{color.by} must set to "split.var"; and vice versa.}

\item{clust.row, clust.column}{Whether to clusters row and/or columns}

\item{add.points}{Whether to plot data}

\item{points.size}{Size for data points}

\item{column.fontsize, row.fontsize, row.title.fontsize, legend.fontsize, legend.title.fontsize, features.fontsize}{Font size for figure and legend}

\item{column.fontface, row.fontface, features.fontface}{Fontface for column/row/feature labels either "plain", "bold", "italic" or "bold.italic" (default: "plain")}

\item{column.names.rotation}{Rotation for column names}

\item{axis.text.hjust, axis.text.vjust}{Horizontal/vertical justification (in [0, 1]) for axis text (default: 0.5)}

\item{colors}{The colors to fill the violin plot when \code{color.by} is set to "median.exp" or "mean.exp"}

\item{random.annotation.col.seed}{Seed to generate random colors for feature annotations}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Main function to plot stacked violin plot
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048
}
\keyword{internal}
