% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedHeatmap.R
\name{SummarizedHeatmap}
\alias{SummarizedHeatmap}
\title{Summarized Heatmap for Seurat or post-integration (pi) object}
\usage{
SummarizedHeatmap(
  object,
  features,
  assay = "RNA",
  split.by = NULL,
  heatmap.cols = NULL,
  clust.row = TRUE,
  clust.column = FALSE,
  row.annotation = NULL,
  row.annotation.cols = NULL,
  column.annotation = NULL,
  column.annotation.cols = NULL,
  show.row.names = TRUE,
  show.column.names = TRUE,
  min.exp = -2,
  max.exp = 2,
  column.names.rotation = 45,
  column.fontsize = 12,
  row.fontsize = 12,
  legend.label.fontsize = 13,
  legend.title.fontsize = 15,
  annotation.name.fontsize = 8,
  heatmap.width = 12,
  heatmap.height = 14,
  random.col.seed = 1,
  ...
)
}
\arguments{
\item{object}{A Seurat or pi object}

\item{features}{A character vector of features}

\item{assay}{Name of the assay to use (default: "RNA")}

\item{split.by}{Factor to split the columns by}

\item{heatmap.cols}{Colors for filling the heatmap}

\item{clust.row, clust.column}{Whether to cluster the rows/columns (default: TRUE)}

\item{row.annotation, column.annotation}{Data.frame for row and column annotation}

\item{row.annotation.cols, column.annotation.cols}{Colors for row/Column annotations; If set to NULL, \code{link[randomcoloR]{distinctColorPalette}} will be used to generate random colors}

\item{show.row.names, show.column.names}{Whether to show row/column names (default: TRUE)}

\item{min.exp, max.exp}{Minimum/maximum scaled expression value to plot (default: -2/2)}

\item{column.names.rotation}{Rotation for column names (default: 90)}

\item{row.fontsize, column.fontsize}{Row/column font size (default: 10/8)}

\item{legend.label.fontsize, legend.title.fontsize, annotation.name.fontsize}{Font sizes for legend label/title and annotation name (default: 10)}

\item{heatmap.width, heatmap.height}{Width and height of heatmap (default: 12/10)}

\item{random.col.seed}{Set seed to control row and column annotation colors (default: 1)}

\item{...}{Extra parameters passed to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object
}
\description{
An internal function to plot heatmap that groups by identity class or additional meta data. Use its wrapper function \code{\link[Ragas]{RunSummarizedHeatmap}} for plotting.
}
\details{
SummarizedHeatmap takes data from the data slot of the Seurat object for the given features and split the cells
and summarize their average expression per identity (or by identity and additional metadata as determined by the split.by argument)
and scaled across all columns.
}
\references{
Hao and Hao et al. (2021). “Integrated analysis of multimodal single-cell data.” Cell. doi:10.1016/j.cell.2021.04.048

Gu, Z. (2016) Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics.
}
\keyword{internal}
